update rename schemas to validators and create project-related validator
