+++ src/renderer/src/lib/validators/device.ts
